+++ browser-switch/src/test/java/com/braintreepayments/api/BrowserSwitchPendingRequestUnitTest.kt
@@ -16,6 +16,7 @@ class BrowserSwitchPendingRequestUnitTest {
         Uri.parse("http://"),
         JSONObject().put("test_key", "test_value"),
         "return-url-scheme",
+        Uri.parse("https://example.com"),
         false
     )