+++ browser-switch/src/test/java/com/braintreepayments/api/BrowserSwitchRequestUnitTest.java
@@ -29,7 +29,7 @@ public class BrowserSwitchRequestUnitTest {
     }
 
     @Test
-    public void fromJson_withoutMetadata() throws JSONException {
+    public void fromJson_withoutMetadata_or_appLinkUri() throws JSONException {
         String json = "{\n" +
                 "  \"requestCode\": 123,\n" +
                 "  \"url\": \"https://example.com\",\n" +
@@ -41,6 +41,7 @@ public class BrowserSwitchRequestUnitTest {
         assertEquals(sut.getRequestCode(), 123);
         assertEquals(sut.getUrl().toString(), "https://example.com");
         assertNull(sut.getMetadata());
+        assertNull(sut.getAppLinkUri());
         assertTrue(sut.getShouldNotifyCancellation());
 
         assertTrue(sut.matchesDeepLinkUrlScheme(Uri.parse("my-return-url-scheme://test")));
@@ -48,7 +49,7 @@ public class BrowserSwitchRequestUnitTest {
     }
 
     @Test
-    public void toJson_withoutMetadata() throws JSONException {
+    public void toJson_withoutMetadata_or_appLinkUri() throws JSONException {
         String json = "{\n" +
                 "  \"requestCode\": 123,\n" +
                 "  \"url\": \"https://example.com\",\n" +
@@ -61,6 +62,7 @@ public class BrowserSwitchRequestUnitTest {
         assertEquals(restored.getRequestCode(), original.getRequestCode());
         assertEquals(restored.getUrl(), original.getUrl());
         assertNull(restored.getMetadata());
+        assertNull(restored.getAppLinkUri());
         assertFalse(restored.getShouldNotifyCancellation());
 
         assertTrue(restored.matchesDeepLinkUrlScheme(Uri.parse("my-return-url-scheme://test")));
@@ -68,12 +70,13 @@ public class BrowserSwitchRequestUnitTest {
     }
 
     @Test
-    public void fromJson_withMetadata() throws JSONException {
+    public void fromJson_withMetadata_and_appLinkUri() throws JSONException {
         String json = "{\n" +
                 "  \"requestCode\": 123,\n" +
                 "  \"url\": \"https://example.com\",\n" +
                 "  \"returnUrlScheme\": \"my-return-url-scheme\",\n" +
                 "  \"shouldNotify\": true,\n" +
+                "  \"appLinkUri\": \"https://example.com\",\n" +
                 "  \"metadata\": {\n" +
                 "    \"testKey\": \"testValue\"" +
                 "  }" +
@@ -82,6 +85,7 @@ public class BrowserSwitchRequestUnitTest {
 
         assertEquals(sut.getRequestCode(), 123);
         assertEquals(sut.getUrl().toString(), "https://example.com");
+        assertEquals(sut.getAppLinkUri().toString(), "https://example.com");
         assertTrue(sut.getShouldNotifyCancellation());
         JSONAssert.assertEquals(sut.getMetadata(), new JSONObject().put("testKey", "testValue"), true);
 
@@ -90,12 +94,13 @@ public class BrowserSwitchRequestUnitTest {
     }
 
     @Test
-    public void toJson_withMetadata() throws JSONException {
+    public void toJson_withMetadata_and_appLinkUri() throws JSONException {
         String json = "{\n" +
                 "  \"requestCode\": 123,\n" +
                 "  \"url\": \"https://example.com\",\n" +
                 "  \"returnUrlScheme\": \"my-return-url-scheme\",\n" +
                 "  \"shouldNotify\": false,\n" +
+                "  \"appLinkUri\": \"https://example.com\",\n" +
                 "  \"metadata\": {\n" +
                 "    \"testKey\": \"testValue\"" +
                 "  }" +
@@ -105,11 +110,33 @@ public class BrowserSwitchRequestUnitTest {
 
         assertEquals(restored.getRequestCode(), original.getRequestCode());
         assertEquals(restored.getUrl(), original.getUrl());
+        assertEquals("https://example.com", restored.getAppLinkUri().toString());
         assertFalse(restored.getShouldNotifyCancellation());
         JSONAssert.assertEquals(restored.getMetadata(), original.getMetadata(), true);
 
         assertTrue(restored.matchesDeepLinkUrlScheme(Uri.parse("my-return-url-scheme://test")));
         assertFalse(restored.matchesDeepLinkUrlScheme(Uri.parse("another-return-url-scheme://test")));
+    }
+
+    @Test
+    public void toJson_for_appLinkUri_without_returnUrlScheme() throws JSONException {
+        String json = "{\n" +
+            "  \"requestCode\": 123,\n" +
+            "  \"url\": \"https://example.com\",\n" +
+            "  \"shouldNotify\": false,\n" +
+            "  \"appLinkUri\": \"https://example.com\",\n" +
+            "  \"metadata\": {\n" +
+            "    \"testKey\": \"testValue\"" +
+            "  }" +
+            "}";
+        BrowserSwitchRequest original = BrowserSwitchRequest.fromJson(json);
+        BrowserSwitchRequest restored = BrowserSwitchRequest.fromJson(original.toJson());
+
+        assertEquals(restored.getRequestCode(), original.getRequestCode());
+        assertEquals(restored.getUrl(), original.getUrl());
+        assertEquals("https://example.com", restored.getAppLinkUri().toString());
+        assertFalse(restored.getShouldNotifyCancellation());
+        JSONAssert.assertEquals(restored.getMetadata(), original.getMetadata(), true);
     }
 
     @Test
@@ -139,5 +166,85 @@ public class BrowserSwitchRequestUnitTest {
                 "}";
         BrowserSwitchRequest request = BrowserSwitchRequest.fromJson(json);
         assertFalse(request.matchesDeepLinkUrlScheme(Uri.parse("not-my-return-url-scheme://example.com")));
+    }
+
+    @Test
+    public void matchesAppLinkUri_whenTheSame_returnsTrue() throws JSONException {
+        String json = "{\n" +
+            "  \"requestCode\": 123,\n" +
+            "  \"url\": \"https://example.com\",\n" +
+            "  \"returnUrlScheme\": \"my-return-url-scheme\",\n" +
+            "  \"appLinkUri\": \"https://example.com\",\n" +
+            "  \"shouldNotify\": false,\n" +
+            "  \"metadata\": {\n" +
+            "    \"testKey\": \"testValue\"" +
+            "  }" +
+            "}";
+        BrowserSwitchRequest request = BrowserSwitchRequest.fromJson(json);
+        assertTrue(request.matchesAppLinkUri(Uri.parse("https://example.com")));
+    }
+
+    @Test
+    public void matchesAppLinkUri_whenTheSame_withQueryParams_returnsTrue() throws JSONException {
+        String json = "{\n" +
+            "  \"requestCode\": 123,\n" +
+            "  \"url\": \"https://example.com\",\n" +
+            "  \"returnUrlScheme\": \"my-return-url-scheme\",\n" +
+            "  \"appLinkUri\": \"https://example.com\",\n" +
+            "  \"shouldNotify\": false,\n" +
+            "  \"metadata\": {\n" +
+            "    \"testKey\": \"testValue\"" +
+            "  }" +
+            "}";
+        BrowserSwitchRequest request = BrowserSwitchRequest.fromJson(json);
+        assertTrue(request.matchesAppLinkUri(Uri.parse("https://example.com?isAppLink=true")));
+    }
+
+    @Test
+    public void matchesAppLinkUri_whenSchemeIsDifferent_returnsFalse() throws JSONException {
+        String json = "{\n" +
+            "  \"requestCode\": 123,\n" +
+            "  \"url\": \"https://example.com\",\n" +
+            "  \"returnUrlScheme\": \"my-return-url-scheme\",\n" +
+            "  \"appLinkUri\": \"http://example.com\",\n" +
+            "  \"shouldNotify\": false,\n" +
+            "  \"metadata\": {\n" +
+            "    \"testKey\": \"testValue\"" +
+            "  }" +
+            "}";
+        BrowserSwitchRequest request = BrowserSwitchRequest.fromJson(json);
+        assertFalse(request.matchesAppLinkUri(Uri.parse("https://example.com")));
+    }
+
+    @Test
+    public void matchesAppLinkUri_whenHostIsDifferent_returnsFalse() throws JSONException {
+        String json = "{\n" +
+            "  \"requestCode\": 123,\n" +
+            "  \"url\": \"https://example.com\",\n" +
+            "  \"returnUrlScheme\": \"my-return-url-scheme\",\n" +
+            "  \"appLinkUri\": \"https://example.com\",\n" +
+            "  \"shouldNotify\": false,\n" +
+            "  \"metadata\": {\n" +
+            "    \"testKey\": \"testValue\"" +
+            "  }" +
+            "}";
+        BrowserSwitchRequest request = BrowserSwitchRequest.fromJson(json);
+        assertFalse(request.matchesAppLinkUri(Uri.parse("https://another-example.com")));
+    }
+
+    @Test
+    public void matchesAppLinkUri_whenNull_returnsFalse() throws JSONException {
+        String json = "{\n" +
+            "  \"requestCode\": 123,\n" +
+            "  \"url\": \"https://example.com\",\n" +
+            "  \"returnUrlScheme\": \"my-return-url-scheme\",\n" +
+            "  \"shouldNotify\": false,\n" +
+            "  \"metadata\": {\n" +
+            "    \"testKey\": \"testValue\"" +
+            "  }" +
+            "}";
+        BrowserSwitchRequest request = BrowserSwitchRequest.fromJson(json);
+        assertNull(request.getAppLinkUri());
+        assertFalse(request.matchesAppLinkUri(Uri.parse("https://example.com")));
     }
 }